requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
pytest==8.3.4
httpx==0.28.1
python-dotenv==1.0.1